Migrate VerifyPhraseWalletScene state management to ViewModel
- Move @State values from VerifyPhraseWalletScene to VerifyPhraseViewModel
- Migrate from ObservableObject to @Observable and @MainActor annotations
- Remove @Published wrappers from ViewModel properties
- Migrate isPresentingAlertMessage from Scene to ViewModel
- Move onImportWallet action method to ViewModel
- Update Scene to use @State private var model pattern instead of @StateObject
- Remove redundant Actions extension from Scene
- Follow ImportWalletScene pattern for state management
